refactor(bpane-host): extract scroll submodule from main.rs
Move all scroll displacement detection, residual analysis, and policy
heuristics into a dedicated scroll/ submodule with clean separation:

- scroll/detect.rs: pixel-based column scroll detection (135 lines)
- scroll/residual.rs: tile-level residual dirty-set analysis (118 lines)
- scroll/policy.rs: ScrollCopy, quantisation, capture cadence (114 lines)
- scroll/tests.rs: all 17 existing scroll tests (222 lines)

Each file stays well under 150 lines. main.rs reduced from 3256 to 2683
lines (-573). All 271 tests pass.
