+++ code/apps/bpane-host/src/main.rs
@@ -12,6 +12,7 @@ mod input;
 mod ipc;
 mod region;
 mod resize;
+mod scroll;
 pub mod tiles;
 mod video_classify;
 
@@ -32,11 +33,7 @@ use bpane_protocol::{
 
 use input::{InputBackend, TestInputBackend};
 
-const CONTENT_SCROLL_SEARCH_MAX_PX: usize = 384;
-const SCROLL_RESIDUAL_FULL_REPAINT_RATIO_DEFAULT: f32 = 0.70;
-const SCROLL_DEFER_REPAIR_MAX_INTERIOR_RATIO: f32 = 0.82;
-const SCROLL_DEFER_REPAIR_MIN_SAVED_RATIO: f32 = 0.20;
-const SCROLL_DEFER_REPAIR_MAX_ROW_SHIFT: i32 = 2;
+// Scroll constants moved to scroll module.
 // MIN_EDITABLE_HINT_*_PX moved to region module.
 
 #[derive(Parser, Debug)]
@@ -68,6 +65,14 @@ use region::{
     expand_tile_bounds, extend_dirty_with_tile_bounds, hash_tile_region,
     point_in_capture_region, region_meets_editable_minimum, region_meets_video_minimum,
     scale_css_px_to_screen_px,
+};
+use scroll::{
+    build_scroll_exposed_strip_emit_coords, build_scroll_residual_emit_coords,
+    can_emit_scroll_copy, content_scroll_search_limit_px, detect_column_scroll,
+    has_scroll_region_split, is_content_tile_in_scroll_region, is_scroll_delta_quantized,
+    next_scroll_active_capture_frames, offset_tile_rect_for_emit, select_capture_frame_interval,
+    select_wheel_trusted_scroll, should_defer_scroll_repair, should_emit_scroll_copy,
+    tile_matches_shifted_prev,
 };
 use video_classify::{
     bbox_center_shift, bbox_iou, compute_tile_motion_features, is_photo_like_tile,
@@ -486,12 +491,12 @@ async fn run_ffmpeg_session(
         const VIDEO_MIN_HOLD_FRAMES: u8 = 12;
         const MIN_SCROLL_DY_PX: i32 = 2;
         const INPUT_MIN_SCROLL_DY_PX: i32 = 4;
-        const MAX_CDP_SCROLL_DY_PX: i64 = CONTENT_SCROLL_SEARCH_MAX_PX as i64;
+        const MAX_CDP_SCROLL_DY_PX: i64 = scroll::CONTENT_SCROLL_SEARCH_MAX_PX as i64;
         const CDP_CONTENT_DY_DIVERGENCE_LOG_PX: i32 = 3;
         const INPUT_SCROLL_MIN_CONFIDENCE: f32 = 0.80;
         const NO_INPUT_SCROLL_MIN_CONFIDENCE: f32 = 0.86;
         const SCROLL_SUPPRESS_VIDEO_FRAMES: u8 = 14;
-        const SCROLL_RESIDUAL_FULL_REPAINT_RATIO: f32 = SCROLL_RESIDUAL_FULL_REPAINT_RATIO_DEFAULT;
+        const SCROLL_RESIDUAL_FULL_REPAINT_RATIO: f32 = scroll::SCROLL_RESIDUAL_FULL_REPAINT_RATIO_DEFAULT;
         // Reduce per-tick bandwidth on sub-tile scroll when residual diff is
         // broad/noisy: send exposed strip while scrolling, then one repair.
         const SCROLL_THIN_MODE_RESIDUAL_RATIO: f32 = 0.14;
@@ -2410,371 +2415,7 @@ async fn handle_control_test(
 
 // Geometry helpers moved to region module.
 
-// TileMotionFeatures, compute_tile_motion_features, bbox_iou,
-// bbox_center_shift, is_photo_like_tile moved to video_classify module.
-
-/// Compute the same tile extraction rect used by the tile emitter when a
-/// vertical grid offset is active.
-fn offset_tile_rect_for_emit(
-    coord: tiles::TileCoord,
-    grid: &tiles::TileGrid,
-    offset_y: u16,
-) -> tiles::Rect {
-    let ts = grid.tile_size;
-    let x = coord.col as u16 * ts;
-    let raw_y = coord.row as i32 * ts as i32 - offset_y as i32;
-    let fb_y = raw_y.max(0) as u16;
-    let fb_end_y = ((raw_y + ts as i32).min(grid.screen_h as i32)).max(0) as u16;
-    let h = fb_end_y.saturating_sub(fb_y);
-    let w = ts.min(grid.screen_w.saturating_sub(x));
-    if w == 0 || h == 0 {
-        return tiles::Rect::new(0, 0, 0, 0);
-    }
-    tiles::Rect::new(x, fb_y, w, h)
-}
-
-/// Returns true when `current` tile pixels equal `previous` shifted by scroll dy.
-/// A mismatch or out-of-bounds mapping marks the tile as residual-dirty.
-fn tile_matches_shifted_prev(
-    current: &[u8],
-    previous: &[u8],
-    stride: usize,
-    screen_h: usize,
-    rect: &tiles::Rect,
-    scroll_dy: i16,
-) -> bool {
-    if rect.w == 0 || rect.h == 0 {
-        return true;
-    }
-    let dy = scroll_dy as i32;
-    let x_bytes = rect.x as usize * 4;
-    let row_bytes = rect.w as usize * 4;
-
-    for row in 0..rect.h as usize {
-        let cy = rect.y as i32 + row as i32;
-        let py = cy + dy;
-        if py < 0 || py >= screen_h as i32 {
-            return false;
-        }
-
-        let curr_off = cy as usize * stride + x_bytes;
-        let prev_off = py as usize * stride + x_bytes;
-        let curr_end = curr_off + row_bytes;
-        let prev_end = prev_off + row_bytes;
-        if curr_end > current.len() || prev_end > previous.len() {
-            return false;
-        }
-        if current[curr_off..curr_end] != previous[prev_off..prev_end] {
-            return false;
-        }
-    }
-
-    true
-}
-
-/// Build the residual dirty set for a trusted vertical scroll frame.
-///
-/// For each emit tile, compare current pixels against previous-frame pixels
-/// shifted by `scroll_dy`; only mismatches are emitted.
-fn build_scroll_residual_emit_coords(
-    current: &[u8],
-    previous: &[u8],
-    stride: usize,
-    grid: &tiles::TileGrid,
-    grid_offset_y: u16,
-    scroll_dy: i16,
-    emit_coords: &[tiles::TileCoord],
-) -> Vec<tiles::TileCoord> {
-    if scroll_dy == 0 || emit_coords.is_empty() {
-        return emit_coords.to_vec();
-    }
-    let screen_h = grid.screen_h as usize;
-    let mut out = Vec::with_capacity(emit_coords.len() / 2);
-
-    for &coord in emit_coords {
-        let rect = offset_tile_rect_for_emit(coord, grid, grid_offset_y);
-        if rect.w == 0 || rect.h == 0 {
-            continue;
-        }
-        if !tile_matches_shifted_prev(current, previous, stride, screen_h, &rect, scroll_dy) {
-            out.push(coord);
-        }
-    }
-
-    out
-}
-
-/// Build the exposed-strip dirty set for a vertical scroll copy.
-///
-/// Returns only tiles that map partially/fully out of previous-frame bounds
-/// when shifted by `scroll_dy` (i.e. newly exposed top/bottom strip).
-fn build_scroll_exposed_strip_emit_coords(
-    grid: &tiles::TileGrid,
-    grid_offset_y: u16,
-    scroll_dy: i16,
-    emit_coords: &[tiles::TileCoord],
-) -> Vec<tiles::TileCoord> {
-    if scroll_dy == 0 || emit_coords.is_empty() {
-        return Vec::new();
-    }
-    let dy = scroll_dy as i32;
-    let screen_h = grid.screen_h as i32;
-    let mut out = Vec::with_capacity((emit_coords.len() / 8).max(1));
-    for &coord in emit_coords {
-        let rect = offset_tile_rect_for_emit(coord, grid, grid_offset_y);
-        if rect.w == 0 || rect.h == 0 {
-            continue;
-        }
-        let shifted_top = rect.y as i32 + dy;
-        let shifted_bottom = rect.y as i32 + rect.h as i32 - 1 + dy;
-        if shifted_top < 0 || shifted_bottom >= screen_h {
-            out.push(coord);
-        }
-    }
-    out
-}
-
-/// Skip speculative canvas shifts when residual analysis already says the frame
-/// is effectively a full repaint. In that case, ScrollCopy only creates a wrong
-/// intermediate image and the tiles immediately overwrite it.
-fn should_emit_scroll_copy(
-    scroll_residual_fallback_full: bool,
-    scroll_saved_tiles: Option<usize>,
-) -> bool {
-    if scroll_residual_fallback_full {
-        return false;
-    }
-    scroll_saved_tiles.unwrap_or(1) > 0
-}
-
-fn content_scroll_search_limit_px(cdp_scroll_dy_px: Option<i16>) -> usize {
-    let cdp_abs = cdp_scroll_dy_px
-        .map(|dy| (dy as i32).unsigned_abs() as usize)
-        .unwrap_or(0);
-    cdp_abs.clamp(256, CONTENT_SCROLL_SEARCH_MAX_PX)
-}
-
-fn select_capture_frame_interval(
-    base_frame_interval: std::time::Duration,
-    scroll_active_frame_interval: std::time::Duration,
-    scroll_active_capture_frames_remaining: u8,
-) -> std::time::Duration {
-    if scroll_active_capture_frames_remaining > 0 {
-        scroll_active_frame_interval.min(base_frame_interval)
-    } else {
-        base_frame_interval
-    }
-}
-
-fn next_scroll_active_capture_frames(
-    scroll_active_capture_frames_remaining: u8,
-    scroll_active_capture_frames: u8,
-    pending_scrolls: i32,
-    strong_scroll_observed: bool,
-    cdp_scroll_observed: bool,
-) -> u8 {
-    if pending_scrolls > 0 || strong_scroll_observed || cdp_scroll_observed {
-        scroll_active_capture_frames
-    } else {
-        scroll_active_capture_frames_remaining.saturating_sub(1)
-    }
-}
-
-fn should_defer_scroll_repair(
-    quantized_scroll_copy: bool,
-    interior_ratio: f32,
-    saved_tiles: usize,
-    potential_tiles: usize,
-    row_shift: i32,
-) -> bool {
-    if !quantized_scroll_copy || potential_tiles == 0 {
-        return false;
-    }
-    if interior_ratio <= SCROLL_RESIDUAL_FULL_REPAINT_RATIO_DEFAULT
-        || interior_ratio > SCROLL_DEFER_REPAIR_MAX_INTERIOR_RATIO
-    {
-        return false;
-    }
-    let saved_ratio = saved_tiles as f32 / potential_tiles as f32;
-    row_shift.abs() <= SCROLL_DEFER_REPAIR_MAX_ROW_SHIFT
-        && saved_ratio >= SCROLL_DEFER_REPAIR_MIN_SAVED_RATIO
-}
-
-fn select_wheel_trusted_scroll(
-    cdp_dy: i16,
-    input_scroll_dir: i32,
-    content_scroll: Option<(i16, f32, bool, Option<f32>)>,
-) -> Option<(i16, f32, &'static str, bool, Option<f32>)> {
-    let cdp_dir = (cdp_dy as i32).signum();
-    let cdp_direction_matches =
-        input_scroll_dir == 0 || cdp_dir == 0 || input_scroll_dir == cdp_dir;
-    if !cdp_direction_matches {
-        return None;
-    }
-
-    if let Some((
-        content_dy,
-        content_confidence,
-        content_direction_matches,
-        content_min_confidence,
-    )) = content_scroll
-    {
-        let content_dir = (content_dy as i32).signum();
-        let directions_compatible = content_direction_matches
-            && (cdp_dir == 0 || content_dir == 0 || cdp_dir == content_dir);
-        if directions_compatible {
-            return Some((
-                content_dy,
-                content_confidence,
-                "content",
-                content_direction_matches,
-                content_min_confidence,
-            ));
-        }
-    }
-
-    Some((cdp_dy, 1.0, "cdp", cdp_direction_matches, None))
-}
-
-fn is_scroll_delta_quantized(scroll_dy: i16, quantum_px: u16) -> bool {
-    if quantum_px == 0 {
-        return true;
-    }
-    let dy = i32::from(scroll_dy).unsigned_abs();
-    dy % u32::from(quantum_px) == 0
-}
-
-fn can_emit_scroll_copy(scroll_dy: i16, quantum_px: u16, tile_size: u16) -> bool {
-    if tile_size == 0 {
-        return false;
-    }
-    let dy = i32::from(scroll_dy).unsigned_abs();
-    dy != 0 && is_scroll_delta_quantized(scroll_dy, quantum_px) && dy % u32::from(tile_size) == 0
-}
-
-fn has_scroll_region_split(
-    region_top: u16,
-    region_bottom: u16,
-    region_right: u16,
-    screen_h: u16,
-    screen_w: u16,
-) -> bool {
-    region_top > 0 || region_bottom < screen_h || region_right < screen_w
-}
-
-fn is_content_tile_in_scroll_region(
-    coord: tiles::TileCoord,
-    tile_size: u16,
-    region_top: u16,
-    region_bottom: u16,
-    region_right: u16,
-) -> bool {
-    let tile_top = coord.row * tile_size;
-    let tile_left = coord.col * tile_size;
-    tile_top >= region_top
-        && tile_top.saturating_add(tile_size) <= region_bottom
-        && tile_left.saturating_add(tile_size) <= region_right
-}
-
-// hash_tile_region moved to region module.
-
-/// Detect vertical scroll displacement by comparing pixel columns between frames.
-///
-/// More reliable than row-hash comparison because vertical columns pass through
-/// diverse content regions (headers, text, footers), reducing false positives
-/// from uniform horizontal bands (white backgrounds).
-///
-/// Samples 5 columns at [10%, 30%, 50%, 70%, 90%] of screen width.
-/// For each candidate dy, counts matching pixels across all columns.
-/// Returns the displacement if ≥3 of 5 columns agree.
-fn detect_column_scroll(
-    current: &[u8],
-    previous: &[u8],
-    stride: usize,
-    width: usize,
-    height: usize,
-    max_search: usize,
-) -> Option<(i32, f32)> {
-    if height < 16
-        || width < 16
-        || current.len() < stride * height
-        || previous.len() < stride * height
-    {
-        return None;
-    }
-
-    let max_dy = max_search.min(height / 2);
-    // Sample 5 columns at different x positions for diversity
-    let sample_xs: Vec<usize> = [10, 30, 50, 70, 90]
-        .iter()
-        .map(|pct| (*pct * width / 100).min(width - 1))
-        .collect();
-
-    let mut best_dy: i32 = 0;
-    let mut best_total_matches: usize = 0;
-    let mut best_agreeing_cols: usize = 0;
-
-    for dy in 1..=max_dy {
-        // Try both positive and negative displacement
-        for sign in [1i32, -1i32] {
-            let signed_dy = dy as i32 * sign;
-            let mut total_matches = 0usize;
-            let mut agreeing_cols = 0usize;
-            let overlap = height - dy;
-
-            for &col_x in &sample_xs {
-                let byte_offset = col_x * 4;
-                let mut col_matches = 0usize;
-
-                for y in 0..overlap {
-                    let (curr_y, prev_y) = if sign > 0 {
-                        (y, y + dy) // content moved up: curr[y] was prev[y+dy]
-                    } else {
-                        (y + dy, y) // content moved down: curr[y+dy] was prev[y]
-                    };
-
-                    let curr_off = curr_y * stride + byte_offset;
-                    let prev_off = prev_y * stride + byte_offset;
-
-                    // Compare RGBA pixel (4 bytes)
-                    if curr_off + 4 <= current.len()
-                        && prev_off + 4 <= previous.len()
-                        && current[curr_off..curr_off + 4] == previous[prev_off..prev_off + 4]
-                    {
-                        col_matches += 1;
-                    }
-                }
-
-                let col_confidence = col_matches as f32 / overlap as f32;
-                if col_confidence > 0.6 {
-                    agreeing_cols += 1;
-                }
-                total_matches += col_matches;
-            }
-
-            if agreeing_cols >= 3 && total_matches > best_total_matches {
-                best_total_matches = total_matches;
-                best_dy = signed_dy;
-                best_agreeing_cols = agreeing_cols;
-            }
-        }
-    }
-
-    if best_dy == 0 || best_agreeing_cols < 3 {
-        return None;
-    }
-
-    let overlap = height - best_dy.unsigned_abs() as usize;
-    let max_possible = overlap * sample_xs.len();
-    let confidence = best_total_matches as f32 / max_possible as f32;
-
-    if confidence > 0.7 {
-        Some((best_dy, confidence))
-    } else {
-        None
-    }
-}
-
+// Scroll functions moved to scroll module.
 // preflight_checks moved to config module.
 
 #[cfg(test)]
@@ -3036,221 +2677,7 @@ mod tests {
         }
     }
 
-    // bbox_iou, tile_motion_features tests moved to video_classify::tests.
-
-    #[test]
-    fn tile_matches_shifted_prev_detects_exposed_edge() {
-        let width = 8usize;
-        let height = 8usize;
-        let stride = width * 4;
-        let mut prev = vec![0u8; stride * height];
-        for y in 0..height {
-            for x in 0..width {
-                let off = y * stride + x * 4;
-                prev[off] = y as u8;
-                prev[off + 1] = y as u8;
-                prev[off + 2] = y as u8;
-                prev[off + 3] = 255;
-            }
-        }
-
-        // current[y] = prev[y+1] for y in 0..height-1 (dy = +1, content moved up)
-        let mut curr = vec![0u8; stride * height];
-        for y in 0..(height - 1) {
-            let dst = y * stride;
-            let src = (y + 1) * stride;
-            curr[dst..dst + stride].copy_from_slice(&prev[src..src + stride]);
-        }
-        for x in 0..width {
-            let off = (height - 1) * stride + x * 4;
-            curr[off] = 0xEE;
-            curr[off + 1] = 0xEE;
-            curr[off + 2] = 0xEE;
-            curr[off + 3] = 255;
-        }
-
-        let interior = tiles::Rect::new(0, 0, width as u16, (height - 1) as u16);
-        assert!(tile_matches_shifted_prev(
-            &curr, &prev, stride, height, &interior, 1
-        ));
-
-        let exposed = tiles::Rect::new(0, (height - 1) as u16, width as u16, 1);
-        assert!(!tile_matches_shifted_prev(
-            &curr, &prev, stride, height, &exposed, 1
-        ));
-    }
-
-    #[test]
-    fn scroll_residual_emit_marks_exposed_rows() {
-        let grid = tiles::TileGrid::new(128, 128, 64);
-        let stride = 128usize * 4;
-        let prev = vec![0u8; stride * 128usize];
-        let curr = vec![0u8; stride * 128usize];
-        let emit_coords: Vec<tiles::TileCoord> = (0..grid.rows)
-            .flat_map(|r| (0..grid.cols).map(move |c| tiles::TileCoord::new(c, r)))
-            .collect();
-
-        let down =
-            build_scroll_residual_emit_coords(&curr, &prev, stride, &grid, 0, 1, &emit_coords);
-        assert!(down.contains(&tiles::TileCoord::new(0, 1)));
-        assert!(down.contains(&tiles::TileCoord::new(1, 1)));
-        assert!(!down.contains(&tiles::TileCoord::new(0, 0)));
-        assert!(!down.contains(&tiles::TileCoord::new(1, 0)));
-
-        let up =
-            build_scroll_residual_emit_coords(&curr, &prev, stride, &grid, 0, -1, &emit_coords);
-        assert!(up.contains(&tiles::TileCoord::new(0, 0)));
-        assert!(up.contains(&tiles::TileCoord::new(1, 0)));
-        assert!(!up.contains(&tiles::TileCoord::new(0, 1)));
-        assert!(!up.contains(&tiles::TileCoord::new(1, 1)));
-    }
-
-    #[test]
-    fn scroll_exposed_strip_marks_only_exposed_rows() {
-        let grid = tiles::TileGrid::new(128, 128, 64);
-        let emit_coords: Vec<tiles::TileCoord> = (0..grid.rows)
-            .flat_map(|r| (0..grid.cols).map(move |c| tiles::TileCoord::new(c, r)))
-            .collect();
-
-        let down = build_scroll_exposed_strip_emit_coords(&grid, 0, 1, &emit_coords);
-        assert!(down.contains(&tiles::TileCoord::new(0, 1)));
-        assert!(down.contains(&tiles::TileCoord::new(1, 1)));
-        assert!(!down.contains(&tiles::TileCoord::new(0, 0)));
-        assert!(!down.contains(&tiles::TileCoord::new(1, 0)));
-
-        let up = build_scroll_exposed_strip_emit_coords(&grid, 0, -1, &emit_coords);
-        assert!(up.contains(&tiles::TileCoord::new(0, 0)));
-        assert!(up.contains(&tiles::TileCoord::new(1, 0)));
-        assert!(!up.contains(&tiles::TileCoord::new(0, 1)));
-        assert!(!up.contains(&tiles::TileCoord::new(1, 1)));
-    }
-
-    #[test]
-    fn scroll_copy_policy_skips_full_repaints() {
-        assert!(!should_emit_scroll_copy(true, Some(12)));
-        assert!(!should_emit_scroll_copy(false, Some(0)));
-        assert!(should_emit_scroll_copy(false, Some(1)));
-        assert!(should_emit_scroll_copy(false, None));
-    }
-
-    #[test]
-    fn content_scroll_search_limit_tracks_large_cdp_deltas() {
-        assert_eq!(content_scroll_search_limit_px(None), 256);
-        assert_eq!(content_scroll_search_limit_px(Some(64)), 256);
-        assert_eq!(content_scroll_search_limit_px(Some(-320)), 320);
-        assert_eq!(content_scroll_search_limit_px(Some(512)), 384);
-    }
-
-    #[test]
-    fn moderate_residual_can_defer_repair_for_small_row_shifts() {
-        assert!(should_defer_scroll_repair(true, 0.76, 48, 160, 2));
-        assert!(should_defer_scroll_repair(true, 0.71, 32, 120, 1));
-    }
-
-    #[test]
-    fn severe_or_low_value_residual_still_forces_full_repaint() {
-        assert!(!should_defer_scroll_repair(true, 0.90, 48, 160, 2));
-        assert!(!should_defer_scroll_repair(true, 0.76, 12, 160, 2));
-        assert!(!should_defer_scroll_repair(true, 0.76, 48, 160, 3));
-        assert!(!should_defer_scroll_repair(false, 0.76, 48, 160, 2));
-    }
-
-    #[test]
-    fn wheel_scroll_prefers_content_when_pixels_confirm_it() {
-        let selected = select_wheel_trusted_scroll(128, 1, Some((64, 0.88, true, Some(0.80))));
-        assert_eq!(selected, Some((64, 0.88, "content", true, Some(0.80))));
-    }
-
-    #[test]
-    fn wheel_scroll_falls_back_to_cdp_when_content_disagrees() {
-        let selected = select_wheel_trusted_scroll(128, 1, Some((-64, 0.90, false, Some(0.86))));
-        assert_eq!(selected, Some((128, 1.0, "cdp", true, None)));
-    }
-
-    #[test]
-    fn wheel_scroll_rejects_cdp_when_input_direction_mismatches() {
-        let selected = select_wheel_trusted_scroll(128, -1, None);
-        assert_eq!(selected, None);
-    }
-
-    #[test]
-    fn scroll_activity_uses_faster_capture_interval() {
-        let base = std::time::Duration::from_millis(100);
-        let active = std::time::Duration::from_millis(50);
-        assert_eq!(select_capture_frame_interval(base, active, 0), base);
-        assert_eq!(select_capture_frame_interval(base, active, 3), active);
-    }
-
-    #[test]
-    fn scroll_activity_refreshes_fast_capture_window() {
-        assert_eq!(next_scroll_active_capture_frames(0, 8, 1, false, false), 8);
-        assert_eq!(next_scroll_active_capture_frames(4, 8, 0, true, false), 8);
-        assert_eq!(next_scroll_active_capture_frames(4, 8, 0, false, true), 8);
-        assert_eq!(next_scroll_active_capture_frames(4, 8, 0, false, false), 3);
-        assert_eq!(next_scroll_active_capture_frames(0, 8, 0, false, false), 0);
-    }
-
-    #[test]
-    fn scroll_delta_quantization_checks_screen_pixels() {
-        assert!(is_scroll_delta_quantized(64, 64));
-        assert!(is_scroll_delta_quantized(-128, 64));
-        assert!(!is_scroll_delta_quantized(96, 64));
-        assert!(is_scroll_delta_quantized(96, 0));
-    }
-
-    #[test]
-    fn scroll_copy_requires_whole_tile_moves() {
-        assert!(can_emit_scroll_copy(64, 64, 64));
-        assert!(can_emit_scroll_copy(-128, 64, 64));
-        assert!(!can_emit_scroll_copy(32, 32, 64));
-        assert!(!can_emit_scroll_copy(96, 64, 64));
-        assert!(!can_emit_scroll_copy(0, 64, 64));
-    }
-
-    #[test]
-    fn scroll_region_split_detects_partial_viewports() {
-        assert!(!has_scroll_region_split(0, 768, 1280, 768, 1280));
-        assert!(has_scroll_region_split(84, 768, 1280, 768, 1280));
-        assert!(has_scroll_region_split(0, 740, 1280, 768, 1280));
-        assert!(has_scroll_region_split(0, 768, 1265, 768, 1280));
-    }
-
-    #[test]
-    fn content_tiles_must_fit_fully_inside_scroll_region() {
-        let ts = 64;
-        let region_top = 84;
-        let region_bottom = 740;
-        let region_right = 1265;
-
-        assert!(!is_content_tile_in_scroll_region(
-            tiles::TileCoord::new(0, 1),
-            ts,
-            region_top,
-            region_bottom,
-            region_right,
-        ));
-        assert!(is_content_tile_in_scroll_region(
-            tiles::TileCoord::new(0, 2),
-            ts,
-            region_top,
-            region_bottom,
-            region_right,
-        ));
-        assert!(!is_content_tile_in_scroll_region(
-            tiles::TileCoord::new(0, 11),
-            ts,
-            region_top,
-            region_bottom,
-            region_right,
-        ));
-        assert!(!is_content_tile_in_scroll_region(
-            tiles::TileCoord::new(19, 2),
-            ts,
-            region_top,
-            region_bottom,
-            region_right,
-        ));
-    }
-
-    // css_pixels_scale_to_framebuffer_pixels moved to region::tests.
+    // Scroll tests moved to scroll::tests.
+    // bbox/motion tests moved to video_classify::tests.
+    // Geometry tests moved to region::tests.
 }
